+++ CHANGES.md
@@ -1,4 +1,9 @@
 ## Changelog ##
+
+- [0.1]:
+    - Teachers can now add submissions that participants can later annotate.
+    - The submission of an AnnoPy can now be viewed on the overview page.
+
 - [0.0.1]:
     - Added plugin template files.
     - Added first capabilities, events and database structure.